requirements.txt
paho-mqtt==1.5.0
requests==2.22.0
rhasspy-hermes==0.2.0
websockets==8.1
